+++ packages/swc-plugin/transform/src/models.rs
@@ -23,6 +23,28 @@ pub enum ImportMember {
     Namespace(ImportNamespaceMember),
 }
 
+impl From<&ImportDefaultSpecifier> for ImportMember {
+    fn from(value: &ImportDefaultSpecifier) -> Self {
+        ImportMember::Default(ImportDefaultMember::new(&value.local))
+    }
+}
+
+impl From<&ImportNamedSpecifier> for ImportMember {
+    fn from(value: &ImportNamedSpecifier) -> Self {
+        if let Some(ModuleExportName::Ident(ident)) = &value.imported {
+            ImportMember::Named(ImportNamedMember::new(&ident, Some(value.local.clone())))
+        } else {
+            ImportMember::Named(ImportNamedMember::new(&value.local, None))
+        }
+    }
+}
+
+impl From<&ImportStarAsSpecifier> for ImportMember {
+    fn from(value: &ImportStarAsSpecifier) -> Self {
+        ImportMember::Namespace(ImportNamespaceMember::new(&value.local))
+    }
+}
+
 #[derive(Debug)]
 pub struct ImportDefaultMember {
     // `import foo from 'foo';`
@@ -121,6 +143,22 @@ impl NamedExportRef {
 pub enum ExportMember {
     Actual(ActualExportMember),
     Binding(BindingExportMember),
+}
+
+impl From<&ExportNamedSpecifier> for ExportMember {
+    fn from(value: &ExportNamedSpecifier) -> Self {
+        match &value.orig {
+            ModuleExportName::Ident(orig_ident) => ExportMember::Actual(ActualExportMember::new(
+                &orig_ident,
+                if let Some(ModuleExportName::Ident(exported_ident)) = &value.exported {
+                    Some(exported_ident.sym.clone())
+                } else {
+                    None
+                },
+            )),
+            ModuleExportName::Str(_) => unimplemented!("TODO"),
+        }
+    }
 }
 
 #[derive(Debug)]
